+++ examples/canopen/can_load_save_config.py
@@ -1,12 +1,11 @@
 import sys
 
-from ingenialink.canopen.network import CanopenNetwork, CAN_DEVICE, CAN_BAUDRATE
+from ingenialink.canopen.network import CAN_BAUDRATE, CAN_DEVICE, CanopenNetwork
+from ingenialink.canopen.servo import CanopenServo
 
 
-def connect_slave():
-    net = CanopenNetwork(device=CAN_DEVICE.IXXAT,
-                         channel=0,
-                         baudrate=CAN_BAUDRATE.Baudrate_1M)
+def connect_slave() -> tuple[CanopenServo, CanopenNetwork]:
+    net = CanopenNetwork(device=CAN_DEVICE.IXXAT, channel=0, baudrate=CAN_BAUDRATE.Baudrate_1M)
 
     servo = net.connect_to_slave(
         target=32,
@@ -15,7 +14,7 @@ def connect_slave():
     return servo, net
 
 
-def load_config_example():
+def load_config_example() -> None:
     """Loads a given configuration file into the drive."""
     servo, net = connect_slave()
     servo.load_configuration(
@@ -30,7 +29,7 @@ def load_config_example():
     net.disconnect_from_slave(servo)
 
 
-def save_config_example():
+def save_config_example() -> None:
     """Saves the drive configuration into a file."""
 
     servo, net = connect_slave()