+++ examples/ethernet/eth_connection.py
@@ -3,7 +3,7 @@ import sys
 from ingenialink.ethernet.network import EthernetNetwork
 
 
-def connection_example():
+def connection_example() -> None:
     net = EthernetNetwork()
     servo = net.connect_to_slave("192.168.2.22",
                                  "../../resources/dictionaries/eve-net-c_eth_1.8.1.xdf",