+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Set RPDOMap values by byte string.
 - EtherCAT connection example script.
+- Typing in examples.
 
 ### Changed
 - The signature of the load_firmware method of EthercatNetwork is changed to add the boot_in_app argument.